move functionality to core.ts
