+++ src/FSPage.ts
@@ -1,15 +1,9 @@
-import { NativeModules } from 'react-native';
 import { generateUUID } from './utils';
-
-const isTurboModuleEnabled = __turboModuleProxy != null;
+import { FullStory } from './core';
 
 type PropertiesWithoutPageName = {
   [key: string]: unknown;
 } & { pageName?: never };
-
-const FullStory = isTurboModuleEnabled
-  ? require('./NativeFullStory').default
-  : NativeModules.FullStory;
 
 const { startPage, endPage, updatePage } = FullStory;
